+++ apps/app/src/@routes/$locale/seller/transaction/$listingId/list.tsx
@@ -7,6 +7,11 @@ export const Route = createFileRoute("/$locale/seller/transaction/$listingId/lis
 	component() {
 		const { listingId } = Route.useParams();
 
-		return <ListingTransactionListPage listingId={listingId} />;
+		return (
+			<ListingTransactionListPage
+				_suspense={"I know"}
+				listingId={listingId}
+			/>
+		);
 	},
 });
