+++ apps/app/src/app/@seller/README.md
@@ -41,4 +41,4 @@ Maps to server API: `/api/seller/*`.
 - Seller transaction action buttons now use the seller transaction mutation SDK surface (`accept`, `dispute`, `reject`, `resolve`) and invalidate `transaction-entry` timeline queries.
 - Seller message detail UI now reads conversation timeline through `@user/transaction-entry`, with `resolved` treated as seller read-only.
 - Seller `TitleContainer` pages that navigate back home now use the shared `@common/nav/BackHomeButton` instead of repeating inline `LinkTo + uiBackButton` wiring.
-- Seller listing-level transaction screen now lives in active scope at `@seller/transaction-listing/*`; its hero banner and item stack now rely on the route-level pending page as the single suspense/loading gate, while local `ui/Item/*` pieces keep seller-local status label mapping, `StatusIcon.tsx`, and folder-local `Preview/*` for the final loaded UI.
+- Seller listing-level transaction screen now lives in active scope at `@seller/transaction-listing/*`; its hero banner is flattened into a single local component, transaction rows keep their local `ui/Item/*` stack, and seller transaction status labeling now lives in `@seller/transaction/toStatusLabel.ts`.
