--- apps/app/src/app/@seller/transaction-listing/ListingTransactionHero/ListingTransactionHero.tsx
@@ -1,12 +1,0 @@
-import type { FC } from "react";
-import { Data } from "./Data";
-
-export namespace ListingTransactionHero {
-	export interface Props extends Data.Props {
-		//
-	}
-}
-
-export const ListingTransactionHero: FC<ListingTransactionHero.Props> = (props) => {
-	return <Data {...props} />;
-};
