--- apps/app/src/app/@seller/transaction-listing/ListingTransactionHero/index.ts
@@ -1,1 +1,0 @@
-export { ListingTransactionHero } from "./ListingTransactionHero";
