+++ apps/app/src/app/@seller/transaction-listing/page/ListingTransactionListPage.tsx
@@ -1,3 +1,4 @@
+import type { MarkSuspense } from "@use-pico/client/type";
 import { Container } from "@use-pico/client/ui/container";
 import { translator } from "@use-pico/common/translator";
 import type { tTransactionQuery } from "@zbav-se.me/sdk/api/seller";
@@ -9,12 +10,13 @@ import { ListingTransactionHero } from "../ListingTransactionHero";
 import { TransactionList } from "../ui/TransactionList";
 
 export namespace ListingTransactionListPage {
-	export interface Props extends TitleContainer.Props {
+	export interface Props extends TitleContainer.Props, MarkSuspense.Props {
 		listingId: string;
 	}
 }
 
 export const ListingTransactionListPage: FC<ListingTransactionListPage.Props> = ({
+	_suspense,
 	listingId,
 	...props
 }) => {
@@ -43,9 +45,13 @@ export const ListingTransactionListPage: FC<ListingTransactionListPage.Props> = 
 			{...props}
 		>
 			<Container>
-				<ListingTransactionHero listingId={listingId} />
+				<ListingTransactionHero
+					_suspense={_suspense}
+					listingId={listingId}
+				/>
 
 				<TransactionList
+					_suspense={_suspense}
 					query={query}
 					ui={{
 						inner: "default",
