+++ apps/app/src/app/@seller/transaction-listing/ui/TransactionList.tsx
@@ -1,3 +1,4 @@
+import type { MarkSuspense } from "@use-pico/client/type";
 import type { Container } from "@use-pico/client/ui/container";
 import type { tTransactionQuery } from "@zbav-se.me/sdk/api/seller";
 import type { FC } from "react";
@@ -5,19 +6,21 @@ import { Item } from "./Item";
 import { TransactionListContainer } from "./TransactionListContainer";
 
 export namespace TransactionList {
-	export interface Props extends Container.Props {
+	export interface Props extends Container.Props, MarkSuspense.Props {
 		query: tTransactionQuery;
 	}
 }
 
-export const TransactionList: FC<TransactionList.Props> = ({ query, ...props }) => {
+export const TransactionList: FC<TransactionList.Props> = ({ _suspense, query, ...props }) => {
 	return (
 		<TransactionListContainer
+			_suspense={_suspense}
 			query={query}
 			renderItem={(transactionId) => (
 				<Item
 					key={transactionId}
 					data-id={transactionId}
+					_suspense={_suspense}
 					transactionId={transactionId}
 				/>
 			)}
