+++ apps/app/src/app/@seller/transaction/~public/toStatusLabel.ts
@@ -1,0 +1,1 @@
+export { toStatusLabel } from "../toStatusLabel";
